pyproject.toml
[project]
name = "ichatbio-ncbi-agent"
version = "1"
requires-python = ">=3.12"
dependencies = [
    "ichatbio-sdk~=0.2.1",
    "openai~=1.79.0",
    "pydantic~=2.11.4",
    "instructor~=1.8.2",
    "pytest~=8.3.5",
    "pytest-asyncio~=1.0.0a1",
    "uvicorn~=0.34.3",
    "python-dotenv>=1.2.1",
    "httpx>=0.28.1",
    "tenacity>=9.1.4",
    "xmltodict>=1.0.4",
    "defusedxml>=0.7.1",
]

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
log_cli = true
log_cli_level = "INFO"
